ext,configs: drops and adds numa-enabled stdlib boards
This change drops all the boards and configs created previously
to simulate disaggregated memory systems as those were not strictly
adhering to the stdlib rules. This new change posts most of the
previously written code to create a new abstract disaggregated
memory board for both ARM and RISCV architectures, which minimizes
code duplication and can be easily extended. Extension suggestions
and tips are now commented thoughout the python files. This is a
proper directory structure to navigate these changes.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>