config,ext: added a arm shared memory board
This change adds a new ARM shared memory board, which allows users to
perform shared memory programs across two independent systems. The board
uses a UIO device to expose a physical memory address range to the OS,
which can be then mmaped by a C program in each of the systems.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>