misc: Refactor out # of RL Time Slots as configurable
Parameters like delays of components like crosspoint,
splitter, merger, etc. are configurable.
The above decide the length of a time slot.
According to the paper, 8 time slots of 92 ps, make
the frequency of the network 1.4 GHz. This makes sense,
as 736 ps time period is roughly 1.4 GHz.
So, given configurable frequencies, the Layer computes
the max range of values and the max packets sent
per window.
Multiple frequencies can be defined to make multiple
Layers.