Warn on Null MessageBuffer
Added a warn message to PerfectSwitch, so that if it attempts to
add an null inport we will throw out a warning.

This buffer being null can cause an error that is difficult to
trace.

In this case, what caused the buffer to be NULL was incorrectly
setting 'network="From"' instead of 'network="To"' when defining
a MessageBuffer in the SLICC files.