[WiP] x86,mem: implementing a permission checker
This change extends a simple memobject to implement a permission
checker. In addition there is a new X86 board and a new cache hierarchy
that checks for memory requests' permissions. It is a WiP cahnge for now
as there needs to be memory accesses going to the memory. This is
planned to be implemented as a memory traffic to model the memory checks
correctly for estimating the slowdown of a secure memory.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>