misc: Add active sources
Based on a parameter, we can either set a raw number or a fraction of the srcs
to be active FOR SOURCE SENDING.
Destinations are all still fully active.
There is also a boolean that can determine if these active sources should be
sequential or if they can be random.
Also, right now I am using a hard-coded seed to make sure the random sources
are the same.
Perhaps might consider setting seed through a parameter later.