x86,simobject,mem: Added a permission checker
This change builds on top of the permission tables for secure memory. In
this commit, a new packet is created and send to the memory to fetch the
permission table from the memory. Whenever there is a permission cache
hit, an event is scheduled with hit latency. In case there is a miss, an
event is scheduled with a higher latency and a new memory packet is
created with req (mostly, not all), and sent to the memory where the
table is located. Caching in implemented within the SimObject and the
memory request makes sure that memory contention is correctly
implemented.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>