mem: HN to add delay for "remote host" hit.
This change allows adding additional delay in home nodes in case
of a hit upstream. By doing this we can hack model a shared
fabric attached memory that is kept coherent with its PoC beyond
the PoC of each host using it. Adding this change allows to
hack model such systems using one ruby system and one protocol.
It also does provide a theoretical upper bound on the performance
of these systems. For each request, if it's a
remote hit (there are sharers only from a different host), the
request should go to the directory of the FAM, more latency.